+++ src/test/circular-types/fs-tree/directory.ts
@@ -1,4 +1,4 @@
-import { CaseStyle } from "../../../lib/_helpers/rename";
+import { CaseStyle } from "../../../lib/case-style";
 import { ArrayType } from "../../../lib/types/array";
 import { DocumentType } from "../../../lib/types/document";
 import { LiteralType } from "../../../lib/types/literal";